Command indicates write operation
This is quite a big change. The idea is to prevent unecessary recalculation of the Detector* vector held by the InstrumentTree by using queries from the Command::execute to determine if a writeable change (something that would result in a new Component*/Detector*) was performed. If nothing of this type is done, we do not need to touch the Detector* vector held.