New benchmark for NullCommand
NullCommand is not writeable, so the Detector* vector is not rebuilt. According to the benchmark tests, this means that the modify using a non-write Command is ~5000X faster than with a write Command, as is done for the root/move benchmark.

I need to consider if the Commands should really operate over Nodes rather than Cow components. The reason is that any command operating on a Component is going to yield a write. However, something like a rename of a Node is an operation we may want to perform that does not involve a recalculation of the Detector* vector.