Avoid Cow copy for Nodes in Instrument
This small change has made a big improvement to the benchmark tests. By avoiding the original COW copy, the Node increment, does not result in a an automatic node copy for many of the performance tests. We are seeing a x10 additional speedup for those operations that involve a modify on a unique instrument.