InstrumentTree must have source + sample
PointSource and PointSample added as new subtypes of PathComponent. PathComponents separately allow for subtypes specify whether they are sources or samples. Future sources/samples may not be treated as points, and this new design allows for that.

Instrument construction now looks for a source and a sample PathComponent, and the index of such components are stored and used for future access.

Performance may have dropped slightly due to these new changes as the InstrumentTree now has to be slightly larger, and there are additional searches (albeit small, given that most instruments should have far < 100 PathComponents). I'm not 100% convinced that the drop in performance I'm seeing is real, as my development machine is not running as quickly as last week.

Next steps are to hook these new features into the SourceSampleDetectorPathFactory and use that in the constructor of the DetectorInfo.