Do not keep PathFactory. Keep product instead.
We are continuously remaking the Paths (just indexes to PathComponents) every time there is a modification. This is unnecessary. Instead lets keep the product of the PathFactories (the paths for L1s and L2s indexes) and use the Factory only on a temporary basis.

These changes have allowed me to keep the L1 and L2 index vectors as shared_pointers, and that therefore makes copying less expensive.

Have yet to inject the PathFactory. That's the next step.