Write then read. Just like Mantid updates.
I'm updating the Mantid performance tests so that what Mantid does can be measured against the prototype. A common schenario is the write then read. i.e. move or rotate something then read the position of it. In data reduction you always perform a read after a write so it seems sensible to add a benchmark for this. Also Mantids write performance is misleading because it's very low-overhead, but offloads all the poor performance to the read. So a comparision that measures a write alone is not as useful.