Rolling out approach from ComponentVisitorTest
This feature is not done yet. Need to look at approach from ComponentVisitorTest.cpp and apply it to the ComponentMappers.cpp. I should then be able to make Vector of Components and serialize it with a vector<PolymporhicSerializer<ComponentMapperFactory> > as is done in the ComponentVisitorTest.cpp. I should then be able to complete the Serialization for the CompositeComponentMapper, because CompositeComponent stores a vector of Components that will need to be serialized.