Sub tree index
Added a mechanism to get the ComponentProxy indexes for all sub-items of a ComponentProxy index. This is important because this is how we are going to enable the cascading type modify behaviour on the DetectorInfo. Prior to this we needed the tree in place so that we could execute Modify and that is why the Instrument could not be immutable.