DetectorInfo read performance
In terms of performance with reading we're interested in positions and rotations. Since those are on the DetectorInfo now. I've moved the benchmark over to that. I've also created new benchmarks for the positionDetector vs position. The postion access is faster because positions for all components are stored, whereas for Detectors, we first need to establish the component index from the detector index first. This should be an O(1) time factor, but better to measure anyway.