Add PathComponentInfo
Each of DetectorInfo and PathComponentInfo will store local geometry information. ComponentInfo can optionally be used to direct both but does not store this information. ComponentInfo would be useful as a director for a general move of an instrument component because it would be able to correctly apply all necessary changes.

Still a work in progress